Finish basics of new paging module
