--- src/memory/paging_new/levels.rs
@@ -1,27 +1,0 @@
-pub trait TableLevel {}
-
-pub struct Level4;
-pub struct Level3;
-pub struct Level2;
-pub struct Level1;
-
-impl TableLevel for Level4 {}
-impl TableLevel for Level3 {}
-impl TableLevel for Level2 {}
-impl TableLevel for Level1 {}
-
-pub trait HierachicalLevel: TableLevel {
-    type NextLevel: TableLevel;
-}
-
-impl HierachicalLevel for Level4 {
-    type NextLevel = Level3;
-}
-
-impl HierachicalLevel for Level3 {
-    type NextLevel = Level2;
-}
-
-impl HierachicalLevel for Level2 {
-    type NextLevel = Level1;
-}
